django-rest/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.5
djangorestframework==3.12.4
idna==3.7
pytz==2024.1
requests==2.32.1
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
